worker/requirements.txt
schedule==0.3.2
PyYAML==3.11
rethinkdb
